HWBackend8/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
pydantic